Add duplicates cleaner logic to pymoors-macros
